% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddlist.R
\name{collect}
\alias{collect}
\title{Collect from Spark}
\usage{
collect(rdd)
}
\arguments{
\item{rdd}{rddlist}
}
\value{
list The contents of rddlist in a local R list
}
\description{
Collects and unserializes the entire rdd from Spark back into local R.
}
\examples{
x <- list(1:10, letters, rnorm(10))
xrdd <- rddlist(sc, x)
x2 <- collect(xrdd)
identical(x, x2)
}

