% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddlist.R
\name{lapply_rdd}
\alias{lapply_rdd}
\title{Apply a Function over an rddlist}
\usage{
lapply_rdd(X, FUN, cache = TRUE)
}
\arguments{
\item{X}{rddlist}

\item{FUN}{function to apply to each element of X}

\item{cache}{logical - Should the resulting RDD be cached in Spark's
memory?}
}
\value{
rddlist A Spark Java Object representing the resulting rddlist
}
\description{
Modeled after \code{lapply} in base R
}
\examples{
x <- list(1:10, letters, rnorm(10))
xrdd <- rddlist(sc, x)
lapply_rdd(xrdd, head)

}

