% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddlist.R
\name{mapply_rdd}
\alias{mapply_rdd}
\title{Apply a Function to multiple rddlist arguments}
\usage{
mapply_rdd(FUN, ..., cache = TRUE)
}
\arguments{
\item{FUN}{function to apply to each element of X}

\item{...}{rddlists of the same length}

\item{cache}{logical - Should the resulting RDD be cached in Spark's
memory?}
}
\value{
rddlist A Spark Java Object representing the resulting rddlist
}
\description{
Modeled after \code{mapply} in base R
}
\examples{
x <- rddlist(sc, list(1:10, letters, rnorm(10)))
y <- rddlist(sc, list(21:30, LETTERS, rnorm(10)))
xy <- mapply_rdd(c, x, y)

}

