% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddlist.R
\name{zip2}
\alias{zip2}
\title{zip 2 rddlists into one}
\usage{
zip2(a, b, a_nested = FALSE, b_nested = FALSE)
}
\description{
This does the actual work for zip_rdd
}
\details{
a_nested = TRUE means that a is already in the form of a nested list with
two layers: [ [a1], [a2], ... , [an] ]
}

