% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddlist.R
\name{zip_rdd}
\alias{zip_rdd}
\title{Zip rdds together}
\usage{
zip_rdd(..., cache = TRUE)
}
\arguments{
\item{...}{rddlists of the same length}

\item{cache}{logical - Should the resulting RDD be cached in Spark's
memory?}
}
\value{
rddlist A Spark Java Object representing the resulting rddlist
}
\description{
For rdds a, b, ... of the same length n this creates an rddlist of length n
where the ith element has the value list(a[[i]], b[[i]], ... )
}
\details{
It always returns a nested list.
}

